Ramon/ramon/requirements.txt
apt-xapian-index==0.49
apturl==0.5.2
asgiref==3.5.0
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-on-heroku==1.1.2
django-paypal==2.0
djangorestframework==3.13.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
geocoder==1.38.1
gpg===1.13.1-unknown
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
IMDbPY==6.8
importlib-metadata==1.5.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.4
oauthlib==3.1.0
olefile==0.46
pandas==1.4.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pycairo==1.16.2
pycrypto==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
ratelim==0.1.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.2
system-service==0.3
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
whitenoise==6.0.0
xkit==0.0.0
youtube-dl==2020.3.24
zipp==1.0.0
